02_FocusedRoom_website/requirements.txt
alembic==1.10.2
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Flask==2.2.5
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
python-dotenv==1.0.0
requests==2.31.0
setuptools==80.9.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
pytest==8.3.4
sendgrid==6.10.0
email-validator==2.1.0
# Gemini AI Integration (MILESTONE 5)
google-generativeai==0.8.3
# Development and CI/CD dependencies
pytest-cov==4.1.0
black==24.3.0
flake8==7.0.0
isort==5.13.2
ruff==0.3.4
mypy==1.9.0
pre-commit==3.6.2
bandit==1.7.8
safety==3.0.1
types-requests==2.31.0.20240406
